requirements.txt
discord.py>=2.3.2
anthropic>=0.17.1
python-dotenv>=1.0.0
httpx>=0.26.0
